[TAN-4889] Refactored email previews
