+++ back/engines/commercial/flag_inappropriate_content/spec/mailers/previews/email_campaigns/inappropriate_content_flagged_mailer_preview.rb
@@ -2,7 +2,7 @@
 
 module EmailCampaigns
   class InappropriateContentFlaggedPreview < ActionMailer::Preview
-    include EmailCampaigns::MailerPreviewRecipient
+    include EmailCampaigns::MailerPreview
 
     def campaign_mail
       flaggable = Idea.first
