+++ back/engines/commercial/idea_assignment/spec/mailers/previews/email_campaigns/idea_assigned_to_you_mailer_preview.rb
@@ -2,7 +2,7 @@
 
 module EmailCampaigns
   class IdeaAssignedToYouMailerPreview < ActionMailer::Preview
-    include EmailCampaigns::MailerPreviewRecipient
+    include EmailCampaigns::MailerPreview
 
     def campaign_mail
       idea = Idea.first
