+++ back/engines/free/email_campaigns/spec/mailers/previews/email_campaigns/comment_deleted_by_admin_mailer_preview.rb
@@ -2,7 +2,7 @@
 
 module EmailCampaigns
   class CommentDeletedByAdminMailerPreview < ActionMailer::Preview
-    include EmailCampaigns::MailerPreviewRecipient
+    include EmailCampaigns::MailerPreview
 
     def campaign_mail
       comment = Comment.first
