+++ back/engines/free/email_campaigns/spec/mailers/previews/email_campaigns/comment_marked_as_spam_mailer_preview.rb
@@ -2,7 +2,7 @@
 
 module EmailCampaigns
   class CommentMarkedAsSpamMailerPreview < ActionMailer::Preview
-    include EmailCampaigns::MailerPreviewRecipient
+    include EmailCampaigns::MailerPreview
 
     def campaign_mail
       comment = Comment.first
