+++ back/engines/free/email_campaigns/spec/mailers/previews/email_campaigns/community_monitor_report_mailer_preview.rb
@@ -2,7 +2,7 @@
 
 module EmailCampaigns
   class CommunityMonitorReportMailerPreview < ActionMailer::Preview
-    include EmailCampaigns::MailerPreviewRecipient
+    include EmailCampaigns::MailerPreview
 
     def campaign_mail
       campaign = EmailCampaigns::Campaigns::CommunityMonitorReport.new
