+++ back/engines/free/email_campaigns/spec/mailers/previews/email_campaigns/cosponsor_of_your_idea_mailer_preview.rb
@@ -2,7 +2,7 @@
 
 module EmailCampaigns
   class CosponsorOfYourIdeaMailerPreview < ActionMailer::Preview
-    include EmailCampaigns::MailerPreviewRecipient
+    include EmailCampaigns::MailerPreview
 
     def campaign_mail
       campaign = EmailCampaigns::Campaigns::CosponsorOfYourIdea.first
