+++ back/engines/free/email_campaigns/spec/mailers/previews/email_campaigns/event_registration_confirmation_mailer_preview.rb
@@ -2,7 +2,7 @@
 
 module EmailCampaigns
   class EventRegistrationConfirmationMailerPreview < ActionMailer::Preview
-    include EmailCampaigns::MailerPreviewRecipient
+    include EmailCampaigns::MailerPreview
 
     def campaign_mail
       campaign.mailer_class
