+++ back/engines/free/email_campaigns/spec/mailers/previews/email_campaigns/idea_marked_as_spam_mailer_preview.rb
@@ -2,7 +2,7 @@
 
 module EmailCampaigns
   class IdeaMarkedAsSpamMailerPreview < ActionMailer::Preview
-    include EmailCampaigns::MailerPreviewRecipient
+    include EmailCampaigns::MailerPreview
 
     def campaign_mail
       idea = Idea.first
