+++ back/engines/free/email_campaigns/spec/mailers/previews/email_campaigns/idea_published_mailer_preview.rb
@@ -2,7 +2,7 @@
 
 module EmailCampaigns
   class IdeaPublishedMailerPreview < ActionMailer::Preview
-    include EmailCampaigns::MailerPreviewRecipient
+    include EmailCampaigns::MailerPreview
 
     def campaign_mail
       idea = Idea.first
