+++ back/engines/free/email_campaigns/spec/mailers/previews/email_campaigns/internal_comment_on_idea_you_commented_internally_on_mailer_preview.rb
@@ -2,7 +2,7 @@
 
 module EmailCampaigns
   class InternalCommentOnIdeaYouCommentedInternallyOnMailerPreview < ActionMailer::Preview
-    include EmailCampaigns::MailerPreviewRecipient
+    include EmailCampaigns::MailerPreview
 
     def campaign_mail
       # TODO: generate commands with campaign#generate_commands method
