+++ back/engines/free/email_campaigns/spec/mailers/previews/email_campaigns/internal_comment_on_unassigned_unmoderated_idea_mailer_preview.rb
@@ -2,7 +2,7 @@
 
 module EmailCampaigns
   class InternalCommentOnUnassignedUnmoderatedIdeaMailerPreview < ActionMailer::Preview
-    include EmailCampaigns::MailerPreviewRecipient
+    include EmailCampaigns::MailerPreview
 
     def campaign_mail
       # TODO: generate commands with campaign#generate_commands method
