+++ back/engines/free/email_campaigns/spec/mailers/previews/email_campaigns/invitation_to_cosponsor_idea_mailer_preview.rb
@@ -2,7 +2,7 @@
 
 module EmailCampaigns
   class InvitationToCosponsorIdeaMailerPreview < ActionMailer::Preview
-    include EmailCampaigns::MailerPreviewRecipient
+    include EmailCampaigns::MailerPreview
 
     def campaign_mail
       campaign = EmailCampaigns::Campaigns::InvitationToCosponsorIdea.first
