+++ back/engines/free/email_campaigns/spec/mailers/previews/email_campaigns/invite_received_mailer_preview.rb
@@ -2,7 +2,7 @@
 
 module EmailCampaigns
   class InviteReceivedMailerPreview < ActionMailer::Preview
-    include EmailCampaigns::MailerPreviewRecipient
+    include EmailCampaigns::MailerPreview
 
     def campaign_mail
       inviter = User.last
