+++ back/engines/free/email_campaigns/spec/mailers/previews/email_campaigns/invite_reminder_mailer_preview.rb
@@ -2,7 +2,7 @@
 
 module EmailCampaigns
   class InviteReminderMailerPreview < ActionMailer::Preview
-    include EmailCampaigns::MailerPreviewRecipient
+    include EmailCampaigns::MailerPreview
 
     def campaign_mail
       inviter = User.last
