+++ back/engines/free/email_campaigns/spec/mailers/previews/email_campaigns/mention_in_official_feedback_mailer_preview.rb
@@ -2,7 +2,7 @@
 
 module EmailCampaigns
   class MentionInOfficialFeedbackMailerPreview < ActionMailer::Preview
-    include EmailCampaigns::MailerPreviewRecipient
+    include EmailCampaigns::MailerPreview
 
     def campaign_mail
       campaign = EmailCampaigns::Campaigns::MentionInOfficialFeedback.first
