+++ back/engines/free/email_campaigns/spec/mailers/previews/email_campaigns/moderator_digest_mailer_preview.rb
@@ -2,7 +2,7 @@
 
 module EmailCampaigns
   class ModeratorDigestMailerPreview < ActionMailer::Preview
-    include EmailCampaigns::MailerPreviewRecipient
+    include EmailCampaigns::MailerPreview
 
     def campaign_mail
       campaign = EmailCampaigns::Campaigns::ModeratorDigest.first
