+++ back/engines/free/email_campaigns/spec/mailers/previews/email_campaigns/native_survey_not_submitted_mailer_preview.rb
@@ -2,7 +2,7 @@
 
 module EmailCampaigns
   class NativeSurveyNotSubmittedMailerPreview < ActionMailer::Preview
-    include EmailCampaigns::MailerPreviewRecipient
+    include EmailCampaigns::MailerPreview
 
     def campaign_mail
       campaign = EmailCampaigns::Campaigns::NativeSurveyNotSubmitted.first
