+++ back/engines/free/email_campaigns/spec/mailers/previews/email_campaigns/new_idea_for_admin_mailer_preview.rb
@@ -2,7 +2,7 @@
 
 module EmailCampaigns
   class NewIdeaForAdminMailerPreview < ActionMailer::Preview
-    include EmailCampaigns::MailerPreviewRecipient
+    include EmailCampaigns::MailerPreview
 
     def campaign_mail
       idea = Idea.published.first
