+++ back/engines/free/email_campaigns/spec/mailers/previews/email_campaigns/project_folder_moderation_rights_received_mailer_preview.rb
@@ -2,7 +2,7 @@
 
 module EmailCampaigns
   class ProjectFolderModerationRightsReceivedMailerPreview < ActionMailer::Preview
-    include EmailCampaigns::MailerPreviewRecipient
+    include EmailCampaigns::MailerPreview
 
     def campaign_mail
       project_folder = ProjectFolders::Folder.first
