+++ back/engines/free/email_campaigns/spec/mailers/previews/email_campaigns/project_phase_upcoming_mailer_preview.rb
@@ -2,7 +2,7 @@
 
 module EmailCampaigns
   class ProjectPhaseUpcomingMailerPreview < ActionMailer::Preview
-    include EmailCampaigns::MailerPreviewRecipient
+    include EmailCampaigns::MailerPreview
 
     def campaign_mail
       campaign = EmailCampaigns::Campaigns::ProjectPhaseUpcoming.first
