+++ back/engines/free/email_campaigns/spec/mailers/previews/email_campaigns/status_change_on_idea_you_follow_mailer_preview.rb
@@ -2,7 +2,7 @@
 
 module EmailCampaigns
   class StatusChangeOnIdeaYouFollowMailerPreview < ActionMailer::Preview
-    include EmailCampaigns::MailerPreviewRecipient
+    include EmailCampaigns::MailerPreview
 
     def campaign_mail
       campaign = EmailCampaigns::Campaigns::StatusChangeOnIdeaYouFollow.first
