+++ back/engines/free/email_campaigns/spec/mailers/previews/email_campaigns/threshold_reached_for_admin_mailer_preview.rb
@@ -2,7 +2,7 @@
 
 module EmailCampaigns
   class ThresholdReachedForAdminMailerPreview < ActionMailer::Preview
-    include EmailCampaigns::MailerPreviewRecipient
+    include EmailCampaigns::MailerPreview
 
     def campaign_mail
       campaign = EmailCampaigns::Campaigns::ThresholdReachedForAdmin.first
