+++ back/engines/free/email_campaigns/spec/mailers/previews/email_campaigns/voting_basket_not_submitted_mailer_preview.rb
@@ -2,7 +2,7 @@
 
 module EmailCampaigns
   class VotingBasketNotSubmittedMailerPreview < ActionMailer::Preview
-    include EmailCampaigns::MailerPreviewRecipient
+    include EmailCampaigns::MailerPreview
 
     def campaign_mail
       project = Project.first
