+++ back/engines/free/email_campaigns/spec/mailers/previews/email_campaigns/voting_basket_submitted_mailer_preview.rb
@@ -2,7 +2,7 @@
 
 module EmailCampaigns
   class VotingBasketSubmittedMailerPreview < ActionMailer::Preview
-    include EmailCampaigns::MailerPreviewRecipient
+    include EmailCampaigns::MailerPreview
 
     def campaign_mail
       project = Project.first
