+++ back/engines/free/email_campaigns/spec/mailers/previews/email_campaigns/voting_results_mailer_preview.rb
@@ -2,7 +2,7 @@
 
 module EmailCampaigns
   class VotingResultsMailerPreview < ActionMailer::Preview
-    include EmailCampaigns::MailerPreviewRecipient
+    include EmailCampaigns::MailerPreview
 
     def campaign_mail
       project = Project.first
