+++ back/engines/free/email_campaigns/spec/mailers/previews/email_campaigns/your_input_in_screening_mailer_preview.rb
@@ -2,7 +2,7 @@
 
 module EmailCampaigns
   class YourInputInScreeningMailerPreview < ActionMailer::Preview
-    include EmailCampaigns::MailerPreviewRecipient
+    include EmailCampaigns::MailerPreview
 
     def campaign_mail
       idea = Idea.first
