+++ back/engines/free/email_campaigns/spec/mailers/previews/reset_password/reset_password_mailer_preview.rb
@@ -2,7 +2,7 @@
 
 module ResetPassword
   class ResetPasswordMailerPreview < ActionMailer::Preview
-    include EmailCampaigns::MailerPreviewRecipient
+    include EmailCampaigns::MailerPreview
 
     def send_confirmation_code
       ::ResetPasswordMailer.with(user: recipient_user, password_reset_url: 'https://demo.stg.govocal.com').send_reset_password
