+++ back/engines/free/email_campaigns/spec/mailers/previews/user_blocking/user_blocked_mailer_preview.rb
@@ -2,7 +2,7 @@
 
 module UserBlocking
   class UserBlockedMailerPreview < ActionMailer::Preview
-    include EmailCampaigns::MailerPreviewRecipient
+    include EmailCampaigns::MailerPreview
 
     def send_user_blocked_email
       recipient_user.update(block_start_at: 5.days.ago, block_end_at: 5.days.from_now, block_reason: 'You breached guideline X')
